% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm-utils.R
\name{add_ppm_data}
\alias{add_ppm_data}
\title{Add weights and responses for poisson point process models}
\usage{
add_ppm_data(ppm_data, presence_var = "presence",
  spat_form = ~longitude + latitude)
}
\arguments{
\item{ppm_data}{Presence only data padded with background points.
See \code{add_quads}.}

\item{presence_var}{The variable in \code{ppm_data} that stores information
on presence/absence.}

\item{spat_form}{RHS formula that indicates names of variables that
store longitude and latitude information (in that order).docu}
}
\description{
Add weights and responses for poisson point process models
}
