% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-utils.R
\name{add_prediction}
\alias{add_prediction}
\alias{add_prediction.gam}
\alias{add_prediction.train}
\title{Add model prediction to data frame}
\usage{
add_prediction(mod, newdata, ...)

\method{add_prediction}{gam}(mod, newdata, ...)

\method{add_prediction}{train}(mod, newdata, ...)
}
\arguments{
\item{mod}{Model of choice.}

\item{newdata}{The data set to which prediction column will be added.}

\item{...}{Further arguments passed to methods.}
}
\description{
Add model prediction to data frame
}
