% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-integration.R
\name{add_probs}
\alias{add_probs}
\title{Create multinomialy weighted observations from raster cells within polygon}
\usage{
add_probs(multipoint_data, folds, mods, pred_grid)
}
\arguments{
\item{multipoint_data}{An augmented data set where polygon-level observations
are represented by multiple points (centroids of 5x5 grid cells within
the polygon)}

\item{folds}{A folds object (list) that contains fold information for
each species}

\item{mods}{A list of models fit to respective \code{folds}.}

\item{pred_grid}{The grid to which \code{models} will be predicted.}
}
\description{
Create multinomialy weighted observations from raster cells within polygon
}
