% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-utils.R
\name{add_spat_dat}
\alias{add_spat_dat}
\alias{get_adm_data}
\title{Add centroid and area polygon information based on GAUL code.}
\usage{
add_spat_dat(data, shp_admin, keep = "ID", by = "GAUL_CODE",
  candidates = c("Admin1_1", "Admin2_1"), id.var = "ID")

get_adm_data(data, candidates = c("Admin1_1", "Admin2_1"),
  spat_form = ~Longitude + Latitude, id_var = "ID")
}
\arguments{
\item{data}{A data set containing GAUL codes (possibly on different
administrative levels; see \code{candidates}).}

\item{shp_admin}{A polygon data frame with \code{by} column containing
GAUL codes}

\item{by}{The column in \code{shp_admin} that contains GAUL code}

\item{candidates}{The columns in \code{data} that contain GAUL codes
on different administrative levels.}
}
\description{
For all \code{candidates} columns in \code{data}, looks up the smallest
polygons not \code{NA} (column names have to be of format \code{*x_y},
if multiple available per row, will pick the one with largest \code{x} and
\code{y}), looks up centroid and area information for respective polygon in
\code{shp_admin} data and adds respective information back to \code{data}.
}
\keyword{internal}
