% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm-utils.R
\name{as_owin}
\alias{as_owin}
\title{Transform object of class extent to object of class owin}
\usage{
as_owin(data, spat_form = ~longitude + latitude, poly = NULL)
}
\description{
Transform object of class extent to object of class owin
}
\keyword{internal}
