% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-utils.R
\name{as_spatial}
\alias{as_spatial}
\alias{add_projection}
\title{Transform data frame to spatial object}
\usage{
as_spatial(data, coord_form = ~longitude + latitude,
  proj = tcruziutils::proj_tcruzi)

add_projection(data, proj = tcruziutils::proj_tcruzi)
}
\arguments{
\item{data}{Data to transform}
}
\description{
Adds coordinates and projection to data set, turning it into an object
of class \code{Spatial*}.
}
