% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb-utils.R
\name{as_xgb_matrix}
\alias{as_xgb_matrix}
\title{Transform data frame to xgb dense matrix}
\usage{
as_xgb_matrix(data, features, label = NULL, weight_col = NULL)
}
\arguments{
\item{data}{A data frame}

\item{features}{A character vector with column names that will be kept in the
dense matrix.}

\item{label}{The name of the column containing the outcome}

\item{weights}{A (optional) numeric vector with sampling weights. Will be
added to the xbg dense matrix using \code{setinfo(..., "weight", weights)}}

\item{scale}{Logical. Should parameters be scaled before they are passed
to \code{\link[xgboost]{xgb.DMatrix}}}
}
\description{
Performs some preprocessing and creates a dense matrix for xgboost using
\code{\link[xgboost]{xgb.DMatrix}}.
}
