% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bbox_tcruzi}
\alias{bbox_tcruzi}
\title{Boundary box for Tcruzi analysis}
\format{An object of class \code{bbox} of length 4.}
\usage{
bbox_tcruzi
}
\description{
Can be used to create boundary box or used to clip spatial objects
}
\keyword{datasets}
