% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{calc_calibration}
\alias{calc_calibration}
\alias{calc_calibration_sp}
\title{Add calibration data to df}
\usage{
calc_calibration(data, truth = "presence", ncut = 50L,
  metrics = c(auc = auc, brier = brier, z_score = z_score, deviance =
  binom_deviance))

calc_calibration_sp(data, truth = "presence", ncut = 50L,
  metrics = c(auc = auc, brier = brier, z_score = z_score, deviance =
  binom_deviance))
}
\arguments{
\item{data}{A data set containing the true (binary) outcome and
a column with predictions.}

\item{truth}{Name of the column containing true outcome}

\item{ncut}{The number of intervals in which predictions will be disretized.
For each interval, the percentage of events (outcome = 1) is calculated.}

\item{metrics}{A named vector of functions (metrics) that will be applied
to outcomes and predictions.}
}
\description{
Add calibration data to df
}
