% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation.R
\name{cv_gam}
\alias{cv_gam}
\title{Cross validation for gam models based on folds 1-4}
\usage{
cv_gam(fold, form, gamma = 1L, family = binomial(),
  mgcv_args = list(select = TRUE, control = list(trace = FALSE), method =
  "fREML", discrete = TRUE), measures = c("auc", "binom_deviance"),
  train_folds = 1:4, eval_fold = 5)
}
\arguments{
\item{fold}{An object returned by \code{\link{get_sp_folds}}.}

\item{gamma}{Increase this beyond 1 to produce smoother models. \code{gamma} multiplies the effective degrees of freedom in the GCV or UBRE/AIC. code{n/gamma} can be viewed as an effective sample size in the GCV score, and this also enables it to be used with REML/ML. Ignored with P-RE/ML or the \code{efs} optimizer. }

\item{family}{
This is a family object specifying the distribution and link to use in
fitting etc (see \code{\link{glm}} and \code{\link{family}}). See 
\code{\link{family.mgcv}} for a full list of what is available, which goes well beyond exponential family.
Note that \code{quasi} families actually result in the use of extended quasi-likelihood 
if \code{method} is set to a RE/ML method (McCullagh and Nelder, 1989, 9.6).
}
}
\description{
Cross validation for gam models based on folds 1-4
}
