% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation.R
\name{get_best_mod_smry}
\alias{get_best_mod_smry}
\alias{get_best_mod}
\alias{get_gp_ls}
\alias{get_spec_smry_df}
\alias{get_test_auc}
\alias{get_best_auc_tab}
\title{Helper functions for cross-validation evaluation}
\usage{
get_best_mod_smry(cv_list, species)

get_best_mod(cv)

get_gp_ls(cv)

get_spec_smry_df(cv_list, folds, settings_df, species)

get_test_auc(cv)

get_best_auc_tab(cv, fold)
}
\arguments{
\item{cv_list}{A list of \code{cv_settings} outputs.}

\item{species}{The species of interest (as character)}

\item{cv}{An object returned by \code{\link{cv_settings}}.}

\item{folds}{A list of objects returned by \code{get_sp_folds}.}

\item{settings_df}{A data frame of modell settings used when
calling \code{cv_settings}.}

\item{fold}{An object returned by \code{\link{get_sp_folds}}.}
}
\description{
Helper functions for cross-validation evaluation
}
