% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb-utils.R
\name{cv_iprev_xgb}
\alias{cv_iprev_xgb}
\title{Run XGB cross validation given parameters}
\usage{
cv_iprev_xgb(data, xgb_params, features, label, deviance_base,
  deviance_base_eval)
}
\arguments{
\item{data}{Data frame containing all necessary variables.}

\item{xgb_params}{A list containing xgboost arguments. Will be used in
\code{do.call(xgboost, xgb_params)}.}

\item{features}{A character vector of covariates in \code{mod_data} to include
in model.}

\item{label}{The name of the outcome column.}

\item{deviance_base}{a list of baseline deviances in each fold to which
the deviance obtained by the xgb will be compared.}

\item{deviance_base_eval}{The same as above, but calculated on evaluation fold
instead of training folds.}
}
\description{
Run XGB cross validation given parameters
}
