% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-evaluation.R
\name{cv_sdm_mod}
\alias{cv_sdm_mod}
\title{Spatial Cross-validation of a TGB based PPM and logistic regression}
\usage{
cv_sdm_mod(fold, form, add_weights = FALSE, select = TRUE,
  measures = c("auc", "z_score", "brier"), family = quasi(link = "log",
  variance = "mu"), method = "fREML", discrete = TRUE,
  mod_fun = "bam", gp_zero = FALSE)
}
\description{
Spatial Cross-validation of a TGB based PPM and logistic regression
}
