% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation.R
\name{cv_setting}
\alias{cv_setting}
\title{Cross validation for one setting}
\usage{
cv_setting(fold, type, GP, candidates, env_grids, gamma = 1:4,
  train_folds = 1:4, eval_fold = 5)
}
\description{
Cross validation for one setting
}
