% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cross-validation.R
\name{cv_settings}
\alias{cv_settings}
\title{Cross validation wrapper for multiple settings}
\usage{
cv_settings(fold, settings_df, candidates, env_grids, train_folds = 1:4,
  eval_fold = 5, gamma = 1:4)
}
\description{
Cross validation wrapper for multiple settings
}
