% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R, R/xgb-utils.R
\name{eval_prev_gam}
\alias{eval_prev_gam}
\alias{eval_prev_xgb}
\title{Weighted binomial deviance}
\usage{
eval_prev_gam(mod, data, use_weight = TRUE)

eval_prev_xgb(mod, data, candidates, label = NULL, weight_col = NULL)
}
\arguments{
\item{mod}{The estimated GAM}

\item{data}{A data frame}

\item{label}{The name of the column containing the outcome}

\item{test}{Test data (a data frame).}
}
\description{
Calculates (weighted) binomial deviance for \code{mgcv::gam} models.
}
