% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{extent_tcruzi}
\alias{extent_tcruzi}
\title{Spatial extent of Tcruzi analysis}
\format{An object of class \code{Extent} of length 1.}
\usage{
extent_tcruzi
}
\description{
Can be used to create boundary box or used to clip spatial objects
}
\keyword{datasets}
