% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thematic-maps.R
\name{get_breaks}
\alias{get_breaks}
\title{Creates a "breaks" sequence for plotting}
\usage{
get_breaks(vec)
}
\arguments{
\item{vec}{A numeric vector.}
}
\description{
Creates a "breaks" sequence for plotting
}
