% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-utils.R
\name{get_file_name}
\alias{get_file_name}
\title{Extract file name without without path and extension.}
\usage{
get_file_name(file_path)
}
\arguments{
\item{file_path}{The full file path.}
}
\description{
Extract file name without without path and extension.
}
