% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-utils.R
\name{get_gaul_data}
\alias{get_gaul_data}
\alias{add_gaul_code}
\title{Extract GAUL CODE from tcruzi data sets}
\usage{
get_gaul_data(data, id_var = "id")

add_gaul_code(data, id_var = "id")
}
\arguments{
\item{data}{Data to transform}

\item{id_var}{Name of the variable that can be used as unique identifier.}
}
\description{
Extract GAUL CODE from tcruzi data sets

Add column with GAUL CODE to data
}
