% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-utils.R
\name{get_grid}
\alias{get_grid}
\title{Load grid from file path, crop extent and potentially mask}
\usage{
get_grid(file_path, extent = extent(-180, 180, -90, 90), shp = NULL)
}
\arguments{
\item{file_path}{The full file path to a grid file}

\item{extent}{The extent of the grid file that should be loaded.
See \code{\link[raster]{crop}}.}

\item{shp}{If a (polygon) shp file provided, it will be used to
mask everything outside the specified polygons (can be slow,
depending on \code{shp} object. See \code{\link[raster]{mask}}.}
}
\description{
Load grid from file path, crop extent and potentially mask
}
