% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-utils.R
\name{get_grids}
\alias{get_grids}
\title{Import multiple grid data sets}
\usage{
get_grids(file_paths, extent = extent(-180, 180, -90, 90), shp = NULL)
}
\arguments{
\item{file_paths}{Full file paths to grid data.}

\item{extent}{The extent of the grid file that should be loaded.
See \code{\link[raster]{crop}}.}

\item{shp}{If a (polygon) shp file provided, it will be used to
mask everything outside the specified polygons (can be slow,
depending on \code{shp} object. See \code{\link[raster]{mask}}.}
}
\description{
Import multiple grid data sets
}
