% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thematic-maps.R
\name{get_pmap_ci}
\alias{get_pmap_ci}
\title{Map with three panels}
\usage{
get_pmap_ci(cv, fold, env_grids, countries, width = 10,
  discrete = FALSE)
}
\arguments{
\item{cv}{An object returned by \code{\link{cv_settings}}.}

\item{fold}{An object returned by \code{\link{get_sp_folds}}.}

\item{env_grids}{A \code{RasterBrick} object containing all
covariate layers used at estimation stage. Maps will reflect
model prediction for the area defined by \code{env_grids}.}

\item{countries}{A \code{SpatialPolygons*} object. Used as
background for the grid layer.}

\item{width}{Extra buffer around the extended hull used for
modelling for which countrie borders will also be shown.
Defaults to 10 degrees.}
}
\description{
Creates figure with three panels, lower and upper CI as well
as prediction (expectation).
}
