% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-background-data.R
\name{get_quads}
\alias{get_quads}
\alias{add_quads}
\title{Add quadrature points to presence only data}
\usage{
get_quads(data, hull, queries, strata = NULL, year_var = "start_year",
  coord_vars = c("longitude", "latitude"), add_raster_layers = TRUE,
  ...)

add_quads(data, ...)
}
\arguments{
\item{data}{The data set that will be augmented with \code{n}
quadrature  points.}

\item{hull}{A polygon object representing the convex hull of the study area.}

\item{queries}{List of search queries.
See \code{\link[mastergrids]{search_files}}.}

\item{strata}{The name of the column in \code{data} that will be used for
stratification. Must be a \code{factor} or \code{character}. After \code{n}
quadrature points are sampled, the \code{strata} column will be added and
filled by sampling values from the empirical distribution of the
\code{strata} variable.}

\item{year_var}{The name of the column in which information on year of
observation is stored.}

\item{coord_vars}{The names of the columns in which the longitude and
latitude information are stored respectively (in that order).}

\item{add_raster_layers}{Logical. Indicates whether covariate inforamtion should be added in this step.}

\item{...}{Parameters passed to \code{get_quads}}
}
\description{
Adds quadrature points (a.k.a. presence background points) to presence only
data, possibly stratified by a factor variable(s).
In order to do so, covariate information has to be available for all
coordinates at which quadrature points are generated (random or
structured/gridded). This is usually the case for environmeltal variables
available as \code{raster*} objects.
}
