% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-evaluation.R
\name{get_sp_folds}
\alias{get_sp_folds}
\title{Add background data and calculate spatial folds}
\usage{
get_sp_folds(data, species, mask, rasterLayer = NULL, k = 5,
  n_blocks = 200, test_fold = k, species_var = "species2",
  coord_form = ~longitude + latitude, width = 5, extent = NULL,
  proj = proj_tcruzi, selection = "random", showBlocks = FALSE,
  progress = FALSE, maskBySpecies = TRUE, ...)
}
\arguments{
\item{data}{A data set with spatial coordinates (but not a \code{Spatial*}
data set).}

\item{species}{The name of the species for which data will be generated.}

\item{rasterLayer}{RasterLayer for visualisation. If provided, this will be used to specify the blocks covering the area.}

\item{k}{Integer value. The number of desired folds for cross-validation. The default is \code{k = 5}.}

\item{n_blocks}{The approximate number of blocks with which the spatial
region will be filled with. Each fold than consists of approximately
\code{n_blocks}/\code{k} blocks.}

\item{test_fold}{Indicates the block numbers that will be associated with
the test data set. Can be a vector. Defaults to the last fold, i.e.
blocks numbered \code{k}.}

\item{species_var}{The name of the column in which names of species are
stored.}

\item{width}{The width by which convex hull around presence locations is
extended (in degrees).}

\item{selection}{Type of assignment of blocks into  folds. Can be \strong{random} (default), \strong{systematic} or \strong{checkerboard} pattern (not working with user-defined blocks).}

\item{showBlocks}{Logical. If TRUE the final blocks with fold numbers will be plotted. A raster layer could be specified
in \code{rasterlayer} argument to be as background.}

\item{progress}{Logical. If TRUE shows a progress bar when \code{numLimit = NULL} in random fold selection.}

\item{maskBySpecies}{Logical. If raster layer is provided and \code{maskBySpecies = TRUE}, the blocks will be
created based on the raster extent, but only those blocks covering species data is kept. The default is \code{TRUE}.}

\item{...}{Further arguments passed to \code{\link[blockCV]{spatialBlock}}}
}
\description{
Add background data and calculate spatial folds
}
