% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range-utils.R
\name{get_variogram}
\alias{get_variogram}
\title{Add an element containing a variogram to a "fold" object}
\usage{
get_variogram(data, form = presence ~ elevation + rainfall +
  reflectance_evi + reflectance_tcb + reflectance_tcw + lst_day + lst_night
  + lst_diff)
}
\description{
Add an element containing a variogram to a "fold" object
}
