% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibration.R
\name{gg_calibration}
\alias{gg_calibration}
\title{Plot calibration curve}
\usage{
gg_calibration(calibration)
}
\arguments{
\item{calibration}{A data frame containing calibration info.
See \code{\link{calc_calibration}}}
}
\description{
Plot calibration curve
}
