% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial-utils.R
\name{grid_points_in_district}
\alias{grid_points_in_district}
\alias{admin_to_multipoint}
\title{Transform rows with polygon location to multi-rows with grid centroids}
\usage{
grid_points_in_district(data, grid, admin_polygons,
  gaul_var = "GAUL_CODE", id_var = "id", spat_form = ~longitude +
  latitude, .progress = interactive())

admin_to_multipoint(data, grid, admin_polygons, gaul_var = "GAUL_CODE",
  id_var = "id", spat_form = ~longitude + latitude,
  .progress = interactive())
}
\description{
Transform rows with polygon location to multi-rows with grid centroids
}
