% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-utils.R
\name{list_files}
\alias{list_files}
\title{List files by path and pattern}
\usage{
list_files(path, pattern, full.names = TRUE, recursive = TRUE)
}
\arguments{
\item{path}{a character vector of full path names; the default
    corresponds to the working directory, \code{\link{getwd}()}.  Tilde
    expansion (see \code{\link{path.expand}}) is performed.  Missing
    values will be ignored.}

\item{pattern}{an optional \link{regular expression}.  Only file names
    which match the regular expression will be returned.}

\item{full.names}{a logical value.  If \code{TRUE}, the directory
    path is prepended to the file names to give a relative file path.
    If \code{FALSE}, the file names (rather than paths) are returned.}

\item{recursive}{logical.  Should the listing recurse into directories?}
}
\description{
A convenience wrapper around \code{list.files}
}
