% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formula-utils.R
\name{make_gam_formula}
\alias{make_gam_formula}
\alias{add_gp}
\title{Create gam formula depending on number of unique values in covariates}
\usage{
make_gam_formula(data, candidates = NULL, lhs = "presence ~",
  bs = "tp", f = "s", long_lat_vars = c("longitude", "latitude"),
  type = c("linear", "smooth", "geospatial"), n_min = 20L, by = NA)

add_gp(formula)
}
\arguments{
\item{data}{The data set that will be modeled. Needs to have a \code{fold}
column indicating cross-validation folds.}

\item{candidates}{Which variables from \code{data} to consider when
building the model formula.}

\item{lhs}{The left hand side of the formula.}

\item{bs}{The type of basis functions to use when
\code{type = "smooth"} or \code{type = "geospatial"}}

\item{f}{The name of the smooth term constructor.}

\item{type}{The type of linear predictor to construct. One of \code{linear},
\code{smooth} or \code{geospatial}.}

\item{n_min}{Only variables that have at least \code{n_min} unique values
in each cross-validation fold will be included in the formula.}

\item{Names}{of longitude and latitude variables in \code{data}.}
}
\description{
Create gam formula depending on number of unique values in covariates

Add Gaussian process term to model formula
}
\keyword{internal}
