% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/string-utils.R
\name{missing_years_grid}
\alias{missing_years_grid}
\title{Report years not available in grid files}
\usage{
missing_years_grid(data, grid_files, year_var = "Start_year")
}
\arguments{
\item{data}{Any data frame (whith year column \code{year_var}).}

\item{grid_files}{A vector of grid files. For each file the year will be
extracted using \code{str_extract_year}.}

\item{year_var}{The name of the variable stroring year information in
\code{data}.}
}
\description{
Report years not available in grid files
}
