% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{p_endemic_zone}
\alias{p_endemic_zone}
\title{Polygon plot of the endemic zone for the T.cruzi analysis}
\format{An object of class \code{tmap} of length 2.}
\usage{
p_endemic_zone
}
\description{
Used as first layer in various thematic maps within the project
}
\keyword{datasets}
