% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppm-utils.R
\name{ppm_fun}
\alias{ppm_fun}
\title{Wrapper function that performs PPM and creates maps}
\usage{
ppm_fun(fold, candidates, env_grids, polygon, queries, gamma,
  ntile = 200, method = "grid")
}
\arguments{
\item{fold}{An object returned by \code{\link{get_sp_folds}}.}

\item{candidates}{Which variables from \code{data} to consider when
building the model formula.}

\item{env_grids}{A \code{RasterBrick} object containing all
covariate layers used at estimation stage. Maps will reflect
model prediction for the area defined by \code{env_grids}.}

\item{polygon}{A \code{SpatialPolygons*} object.}
}
\description{
Wrapper function that performs PPM and creates maps
}
