% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-integration.R
\name{ppm_to_prob_raster}
\alias{ppm_to_prob_raster}
\title{Predict model on grid points within hull, set values outside to 0}
\usage{
ppm_to_prob_raster(fold, mod, pred_grid)
}
\arguments{
\item{fold}{A fold object, containing species information}

\item{mod}{A model object (model fit to data in \code{fold}.}

\item{pred_grid}{The grid to which \code{models} will be predicted.}
}
\description{
Predict model on grid points within hull, set values outside to 0
}
