% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{proj_tcruzi}
\alias{proj_tcruzi}
\title{Spatial projection used for Tcruzi analysis}
\format{An object of class \code{character} of length 1.}
\usage{
proj_tcruzi
}
\description{
Spatial projection used for Tcruzi analysis
}
\keyword{datasets}
