% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add-background-data.R
\name{quad_coords}
\alias{quad_coords}
\title{Generate coordinates of background points within a spatial extent}
\usage{
quad_coords(data, hull, ..., spat_form = ~longitude + latitude)
}
\arguments{
\item{...}{Further arguments passed to \code{\link[spatstat]{quadscheme}}}
}
\description{
Generate coordinates of background points within a spatial extent
}
\keyword{internal}
