% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thematic-maps.R
\name{spat_plot_prep}
\alias{spat_plot_prep}
\title{Plot spatial distribution of data with different geometries}
\usage{
spat_plot_prep(data, shp_admin1, shp_admin2, keep = "prevalence",
  spat_form = ~longitude + latitude)
}
\arguments{
\item{data}{A data set with variables Longitude and Latitude storing
spatial coordinates. Can be specified via \code{spat_form} argument.}

\item{shp_admin1}{A \code{SpatialPolygonsDataFrame} with administrative
districts polygons level 1.}

\item{shp_admin2}{As \code{shp_admin1} but for administrative districts
level 2.}

\item{keep}{The covariate of interest that will be used as intensity
or fill color.}

\item{spat_form}{Specifies longitude and latitude variables in
\code{data}. See \code{\link[sp]{coordinates}}.}
}
\description{
Plots polygon data on administrative districts levels 1 and 2 together with
point data.
}
\keyword{internal}
