% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcruzi_paths}
\alias{tcruzi_paths}
\title{List of paths containing relevant environmental variables}
\format{An object of class \code{list} of length 13.}
\usage{
tcruzi_paths
}
\description{
List of paths containing relevant environmental variables
}
\keyword{datasets}
