% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcruzi_patterns}
\alias{tcruzi_patterns}
\title{List of regular expressions for specific GeoTIFF files}
\format{An object of class \code{character} of length 13.}
\usage{
tcruzi_patterns
}
\description{
List of regular expressions for specific GeoTIFF files
}
\keyword{datasets}
