% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tcruzi_queries}
\alias{tcruzi_queries}
\title{A list of file paths to environmental variables}
\format{An object of class \code{list} of length 32.}
\usage{
tcruzi_queries
}
\description{
A list of file paths to environmental variables
}
\keyword{datasets}
