% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thematic-maps.R
\name{tm_bivariate}
\alias{tm_bivariate}
\alias{tm_bivar_legend}
\alias{tm_bivar_cv}
\title{Bivariate map}
\usage{
tm_bivariate(data, grid, fold, x_name = "Standard Error",
  y_name = "Prediction", palette = stevens.bluered(n = 9),
  alpha = 0.8, main.title = NULL, shp = countries, size_frac = 15,
  name = "prediction")

tm_bivar_legend(palette, hull, size_frac = 15)

tm_bivar_cv(cv, fold, env_grids, discrete = TRUE, ...)
}
\arguments{
\item{data}{A data frame containing colums "prediction" and "se".}

\item{grid}{A place-holder grid/raster.}

\item{fold}{Species information.}

\item{palette}{A bivariate colour palette.}

\item{cv}{Cross validation output for species.}

\item{env_grids}{Environmental covariates}

\item{...}{Further arguments passed to tm_bivariate}

\item{fold}{Fold information for species.}
}
\description{
Bivariate map that simultaneously conveys prediction and uncertainty
}
\keyword{internal}
