% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-evaluation.R
\name{tmap_blocks}
\alias{tmap_blocks}
\title{Visualize spatial CV blocks}
\usage{
tmap_blocks(blocks, mask = NULL, palette_dots = c("steelblue",
  "black"))
}
\arguments{
\item{blocks}{An object returned by \code{get_sp_folds}}

\item{mask}{An object of class \code{SpatialPointsPolygon} (or equivalent
object (e.g. \code{sf})) defining the spatial region in which observations
occurrred}

\item{palette_dots}{A vector of colors for absence and presence points
respectively.}
}
\description{
Given an object returned by \code{get_sp_folds}, creates thematic maps
that visualize the spatial distribution of presence and absence points
as well as the location of blocks associated with different folds.
}
