% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thematic-maps.R
\name{tmap_sdm}
\alias{tmap_sdm}
\title{Map of best model selected via cross validation}
\usage{
tmap_sdm(cv, fold, countries, width = 10)
}
\arguments{
\item{cv}{An object returned by \code{\link{cv_settings}}.}

\item{fold}{An object returned by \code{\link{get_sp_folds}}.}

\item{countries}{A \code{SpatialPolygons*} object. Used as
background for the grid layer.}

\item{width}{Extra buffer around the extended hull used for
modelling for which countrie borders will also be shown.
Defaults to 10 degrees.}
}
\description{
Map of best model selected via cross validation
}
