% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thematic-maps.R
\name{tmap_species}
\alias{tmap_species}
\alias{tmap_presence}
\alias{tm_spec_pred}
\alias{tm_spec_pred_cut}
\title{Thematic maps of species presence}
\usage{
tmap_species(p_endemic_zone, folds, spec)

tmap_presence(p_endemic_zone, folds, spec)

tm_spec_pred(p_endemic_zone, folds, spec, path = "prediction-maps/")

tm_spec_pred_cut(p_endemic_zone, folds, spec, q = 0.95)
}
\description{
Thematic maps of species presence
}
