% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{WhRd}
\alias{WhRd}
\alias{BlYl}
\alias{BlWh}
\alias{WhGr}
\alias{WhBr}
\alias{WhBlue}
\title{Predifined color ramps}
\usage{
WhRd(n)

BlYl(n)

BlWh(n)

WhGr(n)

WhBr(n)

WhBlue(n)
}
\arguments{
\item{n}{Number of colors returned.}
}
\description{
Can be used to define color ramps for spatial plotting of environmental
covariates, e.g. as \code{palette} argument to \code{tmap::tm_*} objects.
}
\section{Functions}{
\itemize{
\item \code{BlYl}: Black Yellow color ramp for nighttime lights

\item \code{BlWh}: Black White color ramp for accessibility

\item \code{WhGr}: White Green color ramp for forests

\item \code{WhBr}: White Brown color ramp

\item \code{WhBlue}: White Blue color ramp for rainfall
}}

\examples{
WhRd(10)
}
\seealso{
\code{\link[grDevices]{colorRampPalette}}
}
