% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb-utils.R
\name{xgb_binom_deviance}
\alias{xgb_binom_deviance}
\title{Binomial deviance evaluation metric for xgboost}
\usage{
xgb_binom_deviance(preds, dtrain)
}
\description{
Can be specified as `eval_metric` in \code{\link[xgboost]{xgboost}}.
Calculates weighted binomial deviance
}
